refactor: 全面使用 u64 以兼容 fast-down

+++ src/sync/executor.rs
@@ -12,7 +12,7 @@ pub struct Executor<A: Action> {
     pub(crate) action: A,
     pub(crate) task_ptrs: ManuallyDrop<Arc<[*const Task]>>, // use pointer to bypass borrow checker
     pub(crate) id: usize,
-    pub(crate) min_chunk_size: usize,
+    pub(crate) min_chunk_size: u64,
     pub(crate) mutex: Arc<Mutex<()>>,
 }
